Add and update files
